requirements.txt
gunicorn==21.2.0
django-db-geventpool==4.0.2
gevent==24.2.1
Django==5.0.3
redis==5.0.1
spotipy==2.23.0
python-decouple==3.8
django-countries==7.5.1
django_taggit==5.0.1
stripe==8.3.0
dropbox-sign==1.3.0
Pillow==9.5.0
psycopg2==2.9.6
django-heroku==0.3.1
sentry-sdk==1.40.5
django-s3file==5.5.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
requests==2.31.0
drf-spectacular==0.27.2
#rest-social-auth==4.2.0